+++ resources/[dg]/dg-ui/[src]/src/types/apps/itemboxes.d.ts
@@ -6,5 +6,6 @@ declare namespace Itemboxes {
   interface Itembox {
     image: string;
     action: string;
+    isLink?: boolean;
   }
 }